+++ src/main/java/project/gsm/yaml/domain/humanities/presentation/HumanitiesController.java
@@ -4,9 +4,10 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+import project.gsm.yaml.domain.humanities.presentation.dto.request.BooksRequest;
 import project.gsm.yaml.domain.humanities.presentation.dto.request.PrizeRequest;
 import project.gsm.yaml.domain.humanities.presentation.dto.response.*;
-import project.gsm.yaml.domain.humanities.service.HumanitiesService;
+import project.gsm.yaml.domain.humanities.service.*;
 
 import javax.validation.Valid;
 
@@ -15,35 +16,40 @@ import javax.validation.Valid;
 @RequiredArgsConstructor
 public class HumanitiesController {
 
-    private final HumanitiesService humanitiesService;
+    private final PrizeService prizeService;
+    private final BookService bookService;
+    private final VolunteerService volunteerService;
+    private HumanitiesCertificateService humanitiesCertificateService;
+    private SportsService sportsService;
+
 
     @GetMapping("/prize")
     public ResponseEntity<PrizeResponse> getPrize() {
-        PrizeResponse prizeResponse = humanitiesService.prizeExecute();
+        PrizeResponse prizeResponse = prizeService.prizeExecute();
         return new ResponseEntity<>(prizeResponse, HttpStatus.OK);
     }
 
     @GetMapping("/books")
     public ResponseEntity<BooksResponse> getBooks(){
-        BooksResponse booksResponse = humanitiesService.booksExecute();
+        BooksResponse booksResponse = bookService.booksExecute();
         return new ResponseEntity<>(booksResponse, HttpStatus.OK);
     }
 
     @GetMapping("/volunteer")
     public ResponseEntity<VolunteersResponse> getVolunteers(){
-        VolunteersResponse volunteersResponse = humanitiesService.volunteersExecute();
+        VolunteersResponse volunteersResponse = volunteerService.volunteersExecute();
         return new ResponseEntity<>(volunteersResponse, HttpStatus.OK);
     }
 
     @GetMapping("/certificate")
     public ResponseEntity<HumanitiesCertificateResponse> getHumanitiesCertificate(){
-        HumanitiesCertificateResponse humanitiesCertificateResponse = humanitiesService.humanitiesCertificateExecute();
+        HumanitiesCertificateResponse humanitiesCertificateResponse = humanitiesCertificateService.humanitiesCertificateExecute();
         return new ResponseEntity<>(humanitiesCertificateResponse, HttpStatus.OK);
     }
 
     @GetMapping("/sports")
     public ResponseEntity<SportsResponse> getSports(){
-        SportsResponse sportsResponse = humanitiesService.sportsExecute();
+        SportsResponse sportsResponse = sportsService.sportsExecute();
         return new ResponseEntity<>(sportsResponse, HttpStatus.OK);
     }
 
@@ -54,6 +60,8 @@ public class HumanitiesController {
     }
 
     @PostMapping("/books")
-    public ResponseEntity<Void> postBooks(@RequestBody @Valid )
+    public ResponseEntity<Void> postBooks(@RequestBody @Valid BooksRequest booksRequest){
+        humanitiesService.
+    }
 
 }
