+++ src/main/java/project/gsm/yaml/domain/humanities/service/HumanitiesCertificateService.java
@@ -1,0 +1,7 @@
+package project.gsm.yaml.domain.humanities.service;
+
+import project.gsm.yaml.domain.humanities.presentation.dto.response.HumanitiesCertificateResponse;
+
+public interface HumanitiesCertificateService {
+    HumanitiesCertificateResponse humanitiesCertificateExecute();
+}
