+++ src/main/java/project/gsm/yaml/domain/humanities/service/PrizeService.java
@@ -1,18 +1,11 @@
 package project.gsm.yaml.domain.humanities.service;
 
+import project.gsm.yaml.domain.humanities.presentation.dto.request.BooksRequest;
 import project.gsm.yaml.domain.humanities.presentation.dto.request.PrizeRequest;
 import project.gsm.yaml.domain.humanities.presentation.dto.response.*;
 
-public interface HumanitiesService {
+public interface PrizeService {
     PrizeResponse prizeExecute();
-
-    BooksResponse booksExecute();
-
-    VolunteersResponse volunteersExecute();
-
-    HumanitiesCertificateResponse humanitiesCertificateExecute();
-
-    SportsResponse sportsExecute();
 
     void postPrizeExecute(PrizeRequest prizeRequest);
 }
