+++ src/main/java/project/gsm/yaml/domain/humanities/service/SportsService.java
@@ -1,0 +1,7 @@
+package project.gsm.yaml.domain.humanities.service;
+
+import project.gsm.yaml.domain.humanities.presentation.dto.response.SportsResponse;
+
+public interface SportsService {
+    SportsResponse sportsExecute();
+}
