+++ src/main/java/project/gsm/yaml/domain/humanities/service/VolunteerService.java
@@ -1,0 +1,7 @@
+package project.gsm.yaml.domain.humanities.service;
+
+import project.gsm.yaml.domain.humanities.presentation.dto.response.VolunteersResponse;
+
+public interface VolunteerService {
+    VolunteersResponse volunteersExecute();
+}
